PLANNER-440: Tests
